+++ ezyplatform-sdk/ezyplatform-common-sdk/src/main/java/org/youngmonkeys/ezyplatform/service/DefaultUserMetaService.java
@@ -22,7 +22,7 @@ import org.youngmonkeys.ezyplatform.entity.UserMeta;
 import org.youngmonkeys.ezyplatform.repo.UserMetaRepository;
 import org.youngmonkeys.ezyplatform.repo.UserMetaTransactionalRepository;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.List;
 import java.util.Map;
@@ -82,10 +82,10 @@ public class DefaultUserMetaService implements UserMetaService {
     }
 
     @Override
-    public BigInteger increaseUserMetaValue(
+    public BigDecimal increaseUserMetaValue(
         long userId,
         String metaKey,
-        BigInteger value
+        BigDecimal value
     ) {
         return userMetaTransactionalRepository.increaseMetaValue(
             userId,
